Create dual Conky version system
- Rename current version to 'conky_bars' (main_bars.lua, sections_bars/)
- Create new version structure (main.lua, sections/)
- Add startup script to choose between versions
- Create placeholder widgets for new version
- Maintain both versions for comparison and development
