+++ sections_bars/common_draw.lua
