+++ sections_bars/temperature.lua
